backend/requirements.txt
asgiref==3.3.0
Django==3.1.3
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
